backends/python/requirements.txt
Django
psycopg2-binary
django-cors-headers
PyJWT
gunicorn
environs
b24pysdk==0.3.0a1
